Move screening information views and containers
New file structure conventions require stateless view names to
end in `Show` or `Form` and are located in `app/javascript/views/`

New file structure conventions require containers to end in `Container`
and are located in `app/javascript/containers/`

[#152464353]
