+++ app/javascript/containers/screenings/ScreeningInformationShowContainer.js
@@ -1,6 +1,6 @@
 import {connect} from 'react-redux'
 import {getErrorsSelector} from 'selectors/screeningInformationSelectors'
-import ScreeningInformationShow from 'screenings/screeningInformation/ScreeningInformationShow'
+import ScreeningInformationShow from 'views/ScreeningInformationShow'
 import {dateTimeFormatter} from 'utils/dateFormatter'
 import COMMUNICATION_METHOD from 'enums/CommunicationMethod'
