+++ spec/javascripts/components/screenings/ScreeningInformationCardViewSpec.jsx
@@ -19,14 +19,12 @@ describe('ScreeningInformationCardView', () => {
 
   it('renders the form container in edit mode', () => {
     const component = renderScreeningInformationCardView({mode: 'edit'})
-      .find('Connect(ScreeningInformationForm)')
-    expect(component.exists()).toEqual(true)
+    expect(component.find('Connect(ScreeningInformationForm)').exists()).toEqual(true)
   })
 
   it('renders the show container in show mode', () => {
     const component = renderScreeningInformationCardView({mode: 'show'})
-      .find('Connect(ScreeningInformationShow)')
-    expect(component.exists()).toEqual(true)
+    expect(component.find('Connect(ScreeningInformationShow)').exists()).toEqual(true)
   })
 })
