--- spec/javascripts/components/screenings/screeningInformation/ScreeningInformationCardViewSpec.jsx
@@ -1,30 +1,0 @@
-import React from 'react'
-import ScreeningInformationCardView from 'screenings/ScreeningInformationCardView'
-import {shallow} from 'enzyme'
-
-describe('ScreeningInformationCardView', () => {
-  function renderScreeningInformationCardView({editable = true, mode}) {
-    const props = {editable, mode}
-    return shallow(<ScreeningInformationCardView {...props} />)
-  }
-
-  it('renders the card header', () => {
-    const header = renderScreeningInformationCardView({
-      mode: 'edit',
-    }).find('ScreeningCardHeader')
-    expect(header.length).toEqual(1)
-    expect(header.props().showEdit).toEqual(false)
-    expect(header.props().title).toEqual('Screening Information')
-  })
-
-  it('renders the form container in edit mode', () => {
-    const component = renderScreeningInformationCardView({mode: 'edit'})
-    expect(component.find('Connect(ScreeningInformationForm)').exists()).toEqual(true)
-  })
-
-  it('renders the show container in show mode', () => {
-    const component = renderScreeningInformationCardView({mode: 'show'})
-    expect(component.find('Connect(ScreeningInformationShow)').exists()).toEqual(true)
-  })
-})
-
